# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Dodgers Insider API",
        "description": "Los Angeles Dodgers real-time stats, schedules, standings, player profiles, and Sabermetrics. Free, powered by MLB official Stats API.",
        "version": "2.1.0"
    },
    "paths": {
...
        "/schedule/recent": {
            "get": {
                "summary": "Recent Schedule",
                "description": "Recent Dodgers game results.",
                "operationId": "recent_schedule_schedule_recent_get",
                "parameters": [
                    {
                        "name": "days",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 30,
                            "minimum": 1,
                            "description": "Number of past days",
                            "default": 7,
                            "title": "Days"
                        },
                        "description": "Number of past days"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/player/{name_or_id}": {
            "get": {
                "summary": "Player Profile",
                "description": "Player profile by name or MLB ID. Returns structured JSON.",
                "operationId": "player_profile_player__name_or_id__get",
                "parameters": [
                    {
                        "name": "name_or_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Name Or Id"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Output language",
                            "enum": [
                                "en",
                                "zh"
                            ],
                            "default": "en",
                            "title": "Lang"
                        },
                        "description": "Output language"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/report": {
            "get": {
                "summary": "Daily Report",
                "description": "Full daily report with schedule, standings, stats, and news.",
                "operationId": "daily_report_report_get",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "Output language",
                            "enum": [
                                "en",
                                "zh"
                            ],
                            "default": "en",
                            "title": "Lang"
                        },
                        "description": "Output language"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/news": {
            "get": {
                "summary": "News",
                "description": "Latest Dodgers news from MLB.com RSS.",
                "operationId": "news_news_get",
                "parameters": [
                    {
                        "name": "count",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 20,
                            "minimum": 1,
                            "description": "Number of news items",
                            "default": 5,
                            "title": "Count"
                        },
                        "description": "Number of news items"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}